bug hunt
chrome doesn't know unsafewindow without a grant-attribute unequal none.
it also works with no grant, but current userscript standard is grant.

Additionally, there is a bug where the script is stopping when there is
some activity-entry where a username-class is included but is not
containing a name, e.g. when submitting premium content.
fix for this was to check if u is null at two points.